fixed some errors that were preventing deployment.

+++ server/src/seeds/cleanDb.ts
@@ -3,14 +3,25 @@ import db from '../config/connection.js';
 
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
+    const model = models[modelName];
+
+    if (!model) {
+      throw new Error(`Model ${modelName} does not exist.`);
+    }
+
+    if (!model?.db?.db) {
+      throw new Error('Database connection not found');
+    }
+
+    let modelExists = await model.db.db.listCollections({
       name: collectionName
-    }).toArray()
+    }).toArray();
 
     if (modelExists.length) {
       await db.dropCollection(collectionName);
     }
   } catch (err) {
+    console.error("Error in cleanDb function:", err);
     throw err;
   }
-}
+};
